refactor file upload class

+++ eletric-bills-backend/src/domain/useCases/create-invoice/CreateInvoiceUseCase.ts
@@ -1,15 +1,13 @@
 import { container } from "tsyringe";
 import { deleteFile } from "../../../utils/file";
-import { uploadFile } from "../../../utils/upload-file";
 import { CustomerService } from "./services/CustomerService";
 import { EletricityInvoiceParser } from "./services/EletricityInvoiceParserService";
 import { InvoiceService } from "./services/InvoiceService";
+import FileManager from "../../infra/file/FileManager";
 class CreateInvoicesUseCase {
   async execute(invoice: Express.Multer.File): Promise<void> {
     if (invoice) {
-      const { url, path } = await uploadFile({
-        file: invoice,
-      });
+      const { url, path } = await FileManager.uploadFile(invoice);
       const customerService = container.resolve(CustomerService);
       const invoiceService = container.resolve(InvoiceService);
 
@@ -20,10 +18,11 @@ class CreateInvoicesUseCase {
       await invoiceService.storeNewInvoice({
         ...invoiceParameters,
         customerId,
-        invoiceUrl: url,
+        invoiceUrl: url || path,
       });
-
-      await deleteFile(path);
+      if (url) {
+        await deleteFile(path);
+      }
     }
   }
 }
